Consolidate graphical editor regression tests
