+++ docassemble/ALWeaver/test_docassemble_compat.py
@@ -1,3 +1,5 @@
+# do not pre load
+
 import os
 from contextlib import nullcontext
 from pathlib import Path
