+++ docassemble/ALWeaver/test_editor_api.py
@@ -1,3 +1,5 @@
+# do not pre load
+
 from io import BytesIO
 from contextlib import nullcontext
 from pathlib import Path
