+++ docassemble/ALWeaver/test_editor_metadata_safety.py
@@ -1,3 +1,5 @@
+# do not pre load
+
 from pathlib import Path
 import unittest
