+++ docassemble/ALWeaver/test_editor_patch_api.py
@@ -1,3 +1,5 @@
+# do not pre load
+
 import unittest
 from unittest.mock import patch
 from pathlib import Path
